+++ cfripper/config/filter.py
@@ -2,28 +2,30 @@ import logging
 import re
 from typing import Any, Callable, Dict, List, Optional, Set, Union
 
-from pydantic import BaseModel, validator
+from pydantic import BaseModel, field_validator
 from pydash.objects import get
 
 from cfripper.model.enums import RuleMode, RuleRisk
 
-VALID_FUNCTIONS = [
+VALID_FUNCTIONS = {
+    "and",
+    "empty",
     "eq",
-    "ne",
-    "lt",
-    "gt",
-    "le",
+    "exists",
     "ge",
+    "gt",
+    "in",
+    "le",
+    "lt",
+    "ne",
     "not",
     "or",
-    "and",
-    "in",
+    "ref",
     "regex",
     "regex:ignorecase",
-    "exists",
-    "empty",
-    "ref",
-]
+    "set",
+    "sorted",
+}
 
 logger = logging.getLogger(__file__)
 
@@ -39,22 +41,37 @@ def get_implemented_filter_function(function_name: str, debug: bool) -> Callable
 
         return wrap
 
+    def single_param_resolver(f):
+        def wrap(*args, **kwargs):
+            calculated_parameters = [arg(kwargs) for arg in args]
+            if len(calculated_parameters) == 1 and isinstance(calculated_parameters[0], (dict, set)):
+                result = f(*calculated_parameters, **kwargs)
+            else:
+                result = f(calculated_parameters, **kwargs)
+            if debug:
+                logger.debug(f"{function_name}({', '.join(str(x) for x in calculated_parameters)}) -> {result}")
+            return result
+
+        return wrap
+
     implemented_filter_functions = {
+        "and": lambda *args, **kwargs: all(arg(kwargs) for arg in args),
+        "empty": param_resolver(lambda *args, **kwargs: len(args) == 0),
         "eq": param_resolver(lambda a, b, **kwargs: a == b),
-        "ne": param_resolver(lambda a, b, **kwargs: a != b),
-        "lt": param_resolver(lambda a, b, **kwargs: a < b),
-        "gt": param_resolver(lambda a, b, **kwargs: a > b),
-        "le": param_resolver(lambda a, b, **kwargs: a <= b),
+        "exists": param_resolver(lambda a, **kwargs: a is not None),
         "ge": param_resolver(lambda a, b, **kwargs: a >= b),
+        "gt": param_resolver(lambda a, b, **kwargs: a > b),
+        "in": param_resolver(lambda a, b, **kwargs: a in b),
+        "le": param_resolver(lambda a, b, **kwargs: a <= b),
+        "lt": param_resolver(lambda a, b, **kwargs: a < b),
+        "ne": param_resolver(lambda a, b, **kwargs: a != b),
         "not": param_resolver(lambda a, **kwargs: not a),
         "or": lambda *args, **kwargs: any(arg(kwargs) for arg in args),
-        "and": lambda *args, **kwargs: all(arg(kwargs) for arg in args),
-        "in": param_resolver(lambda a, b, **kwargs: a in b),
+        "ref": param_resolver(lambda param_name, **kwargs: get(kwargs, param_name)),
         "regex": param_resolver(lambda *args, **kwargs: bool(re.match(*args))),
         "regex:ignorecase": param_resolver(lambda *args, **kwargs: bool(re.match(*args, re.IGNORECASE))),
-        "exists": param_resolver(lambda a, **kwargs: a is not None),
-        "empty": param_resolver(lambda *args, **kwargs: len(args) == 0),
-        "ref": param_resolver(lambda param_name, **kwargs: get(kwargs, param_name)),
+        "set": single_param_resolver(lambda *args, **kwargs: set(*args)),
+        "sorted": single_param_resolver(lambda *args, **kwargs: sorted(*args)),
     }
     return implemented_filter_functions[function_name]
 
@@ -83,9 +100,10 @@ class Filter(BaseModel):
     risk_value: Optional[RuleRisk] = None
     rules: Set[str] = None
 
-    @validator("eval", pre=True)
+    @field_validator("eval", mode="before")
+    @classmethod
     def set_eval(cls, eval, values):
-        return build_evaluator(eval, values["debug"])
+        return build_evaluator(eval, values.data["debug"])
 
     def __call__(self, **kwargs):
         if self.debug: