+++ cfripper/rules/__init__.py
@@ -23,6 +23,7 @@ from cfripper.rules.kms_key_wildcard_principal import KMSKeyWildcardPrincipalRul
 from cfripper.rules.managed_policy_on_user import ManagedPolicyOnUserRule
 from cfripper.rules.policy_on_user import PolicyOnUserRule
 from cfripper.rules.privilege_escalation import PrivilegeEscalationRule
+from cfripper.rules.public_elb_checker_rule import PublicELBCheckerRule
 from cfripper.rules.rds_security_group import RDSSecurityGroupIngressOpenToWorldRule
 from cfripper.rules.s3_bucket_policy import S3BucketPolicyPrincipalRule
 from cfripper.rules.s3_lifecycle_configuration import S3LifecycleConfigurationRule
@@ -38,6 +39,8 @@ from cfripper.rules.sqs_queue_policy import (
     SQSQueuePolicyNotPrincipalRule,
     SQSQueuePolicyPublicRule,
 )
+from cfripper.rules.stack_name_matches_regex import StackNameMatchesRegexRule
+from cfripper.rules.storage_encrypted_rule import StorageEncryptedRule
 from cfripper.rules.wildcard_policies import (
     GenericResourceWildcardPolicyRule,
     S3BucketPolicyWildcardActionRule,
@@ -78,6 +81,7 @@ DEFAULT_RULES = {
         PartialWildcardPrincipalRule,
         PolicyOnUserRule,
         PrivilegeEscalationRule,
+        PublicELBCheckerRule,
         RDSSecurityGroupIngressOpenToWorldRule,
         S3BucketPolicyPrincipalRule,
         S3LifecycleConfigurationRule,
@@ -87,6 +91,7 @@ DEFAULT_RULES = {
         S3BucketPublicReadAclRule,
         S3CrossAccountTrustRule,
         S3ObjectVersioningRule,
+        StorageEncryptedRule,
         SNSTopicDangerousPolicyActionsRule,
         SNSTopicPolicyNotPrincipalRule,
         SNSTopicPolicyWildcardActionRule,
@@ -94,6 +99,7 @@ DEFAULT_RULES = {
         SQSQueuePolicyNotPrincipalRule,
         SQSQueuePolicyPublicRule,
         SQSQueuePolicyWildcardActionRule,
+        StackNameMatchesRegexRule,
         WildcardResourceRule,
     )
 }